Task_1_2_3
I ended up in the middle of task 3, I have the DB layer prepped, then I was doing the jwt token solution and I got stuck. But not on the jwt token. But on a bug in index.js. The deadline for the assignment is tomorrow. So I'm submitting my assignment.